% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance_testing_mtbf.R
\name{acceptance_testing_mtbf}
\alias{acceptance_testing_mtbf}
\title{Reliability acceptance test planning app}
\usage{
acceptance_testing_mtbf(theme = "flatly", width = "100\%",
  storyteller = F, height = if (storyteller) "800px" else "600px",
  css = NULL, more.opts = list(NA), ...)
}
\arguments{
\item{theme}{Character string naming a color theme bootswatch color theme. Must be one of the themes that can be used in code{shinythemes::shinytheme()}}

\item{width}{Width of the printed app. Used for code{rmd = TRUE}, otherwise ignored}

\item{storyteller}{Is this a storyteller app?}

\item{height}{Height of the printed app. Used for code{rmd = TRUE}, otherwise ignored}

\item{css}{Path to a custom css file. If code{NULL} the default css file is used}

\item{more.opts}{Additional options to be passed to the app (see Details)}

\item{...}{Additional arguments passed to code{shiny::runApp()}}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, setting code{pub = TRUE} prevents calls to code{install.packages}. Calls to code{install.packages} should not be included within an app and will result in an error.
}

