% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_css.R
\name{add_css}
\alias{add_css}
\title{Dynamically add css rules to an app}
\usage{
add_css(main.css = getShinyOption("CSS"), user.css = getShinyOption("css"),
  name = getShinyOption("appName"))
}
\arguments{
\item{main.css}{The main CSS file}

\item{user.css}{A user specified CSS file}

\item{name}{The name of the app}
}
\description{
Dynamically add css rules to an app
}
