% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_logo}
\alias{add_logo}
\title{Add a logo to the footer of a navbarPage app}
\usage{
add_logo()
}
\value{
A fontAwesome icon or an image printed in the footer of a
        navbarPage app
}
\description{
Add a logo to the footer of a navbarPage app
}
