% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_package.R
\name{add_packages}
\alias{add_packages}
\title{Install and load an R package}
\usage{
add_packages(pkg = NULL, repo = NULL, pub = FALSE)
}
\arguments{
\item{pkg}{\code{character} Name of a package to be installed/loaded}

\item{repo}{\code{character} Name of the repository from which the package should be installed.}

\item{pub}{\code{logical} variable indicating whether the app be published (see details)}
}
\value{
A printed shiny app
}
\description{
Install and load an R package
}
\details{
If \code{repo = NULL} the package will be installed from the CRAN.  Otherwise, \code{repo} is a character string that referring to the GitHub account in which the package is located.

  When publishing apps on shinyapps.io or shinyServer, attempting to \code{install.packages} will result in an error.  Calls to \code{install.packages} should not be included within an app.
}
