% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rmd.R, R/add_rmd[Conflict].R
\name{add_rmd}
\alias{add_rmd}
\alias{add_rmd}
\title{Add an rmarkdown file to an app}
\usage{
add_rmd(rmd, path)
<<<<<<< HEAD

add_rmd(rmd, path)
=======
>>>>>>> 8f0633dea5da799cab8e03fc56c6b099e85dc324
}
\arguments{
\item{rmd}{\code{character} Name of an rmarkdown file saved in the app directory}

\item{path}{Path to a file outside of the app directory}
<<<<<<< HEAD

\item{rmd}{\code{character} Name of an rmarkdown file saved in the app directory}

\item{path}{Path to a file outside of the app directory}
=======
>>>>>>> 8f0633dea5da799cab8e03fc56c6b099e85dc324
}
\description{
Run inline and stand-alone code chunks and include results
as part of a shiny app. Include \code{LaTeX}-typeset equations with MathJax

Run inline and stand-alone code chunks and include results
as part of a shiny app. Include \code{LaTeX}-typeset equations with MathJax
}
\examples{
\dontrun{
# see examples in add_server and add_ui documentation
}
\dontrun{
# see examples in add_server and add_ui documentation
}
}
\seealso{
\code{\link{add_server} \link{add_ui}}

\code{\link{add_server} \link{add_ui}}
}
