% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_css.R
\name{add_theme}
\alias{add_theme}
\title{Dynamically add a theme to an app}
\usage{
add_theme(theme = NULL)
}
\arguments{
\item{theme}{\code{character} A bootstrap theme provided by shinythemes::shinytheme}
}
\description{
Dynamically add a theme to an app
}
