% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appExists.R
\name{appExists}
\alias{appExists}
\title{Determine if an app exists in a package}
\usage{
appExists(lib.list, app)
}
\arguments{
\item{what}{A \code{list} returned from \code{teachingApps:::what_are_you}}
}
\description{
Determines whether the package in the library or in a local directory
}
