% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appFile.R
\name{appFile}
\alias{appFile}
\title{Function Title}
\usage{
appFile(app, file, pkg = "teachingApps")
}
\arguments{
\item{app}{The app directory in which the file is located}

\item{file}{The name of the desired file}

\item{pkg}{The name of the package in which to search}
}
\description{
Function Title
}
\details{
This function assumes that the apps are stored in a package directory called 'apps'
}

