% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign.shiny.opts.R
\name{assign.shiny.opts}
\alias{assign.shiny.opts}
\title{Pass objects and customization options to apps}
\usage{
assign.shiny.opts(opts, dir, theme = "flatly", icon = "fa fa-github")
}
\arguments{
\item{opts}{\code{list} Additional options or objects to pass to a shiny app}

\item{dir}{\code{character} Path to the directory containing ui.R and server.R}

\item{theme}{\code{character} Bootswatch color theme (used by \code{shinythemes::shinytheme})}

\item{icon}{\code{character} Name of fontAwesome icon printed at the bottom of an app}
}
\value{
A list of options set with \code{shinyOptions}
}
\description{
Provides a general method pass arguments to shiny apps
and ensure the options carry through to app deployment.
}
\details{
Shiny apps are not functions, therefore customization options cannot be
passed to a shiny app as simply as argument are passed between functions.  Further,
the manner in which objects are loaded prior to deploying an app differ if the app
will be published as a stand-alone or embedded within an rmarkdown docmuent.

Assigning objects to shiny options ensures that these values are passed to a shiny
app much the same way as arguments are passed to functions.
}
