% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_histogram.R
\name{basic_histogram}
\alias{basic_histogram}
\title{Simple base R histogram app}
\usage{
basic_histogram(theme = "flatly", width = "100\%", height = "800px",
  more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Description
}
\examples{
\dontrun{
appy(theme = 'spacelab', height = '600px')
}
}
\seealso{
Other shinyAce apps: \code{\link{ace_very_normal}},
  \code{\link{cdf_demo}}
}
