% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bathtub_hazard.R
\name{bathtub_hazard}
\alias{bathtub_hazard}
\title{Visualize hazard function plots with various shapes}
\usage{
bathtub_hazard(theme = "flatly", storyteller = F, width = "100\%",
  height = if (storyteller) "800px" else "600px", icon = "fa fa-github",
  more.opts = list(NA), ...)
}
\arguments{
\item{theme}{\code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{storyteller}{\code{logical} Is this a storyteller app?}

\item{width}{\code{character} Width of the printed app}

\item{height}{\code{character} Height of the printed app}

\item{icon}{\code{character} Name of fontAwesome icon printed at the bottom of an app}

\item{more.opts}{\code{list} Additional options/objects passed to the app (see Details)}

\item{...}{\code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Create uni- and bi-modal hazard function plots
}
\details{
This app displays the hazard function for a 
    system with two independent failure modes.  System failure
    occurs when either of the arranged in series. 
    changes for various parameter values. effect Generates hazard functions plots for a 
    system with two-indeposterior samples by filtering the prior
}
\examples{
\dontrun{
appy(theme = 'spacelab', height = '600px')
}
}
\seealso{
Other statistics apps: \code{\link{bayesian_basic}},
  \code{\link{central_limit}}, \code{\link{complex_ideas}}
}
