% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bent_probplot.R, R/harry_potter.R
\name{bent_probplot}
\alias{bent_probplot}
\alias{bent_probplot}
\title{SMRD non-linear probability plot app}
\usage{
bent_probplot(theme = "flatly", width = "100\%", height = "800px",
  more.opts = list(NA), ...)

bent_probplot(theme = "flatly", width = "100\%", height = "800px",
  more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}

\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app

A printed shiny app
}
\description{
Description

Description
}
\examples{
\dontrun{
appy(theme = 'spacelab', height = '600px')
}
\dontrun{
appy(theme = 'spacelab', height = '600px')
}
}
