% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/berkson_interval.R
\name{berkson_interval}
\alias{berkson_interval}
\title{Function Title}
\usage{
berkson_interval(pub = FALSE, theme = "flatly")
}
\arguments{
\item{pub}{Will this app be published? (see details)}

\item{theme}{Character string naming a color theme bootswatch color theme. Must be one of the themes that can be used in code{shinythemes::shinytheme()}}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, set code{pub = TRUE} to prevent calls to code{install.packages}. Calls to code{install.packages} should not be included within an app and will result in an error.
}

