% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_rows.R
\name{clean_rows}
\alias{clean_rows}
\title{Subset data using column values}
\usage{
clean_rows(data, theme = "flatly", colorBy = NULL, width = "100\%",
  height = "600px", css = NULL)
}
\arguments{
\item{data}{A data set}

\item{theme}{\code{character} A bootswatch theme provided to \code{shinythemes::shinytheme}}

\item{colorBy}{\code{character} Column by which the \code{parcoords} plot should be colored}

\item{width}{\code{character} Width of the gadget (in valid css units)}

\item{height}{\code{character} Height of the gadget (in valid css units)}

\item{css}{\code{character} Path to a custom css file}

\item{rownames}{\code{logical} Should rownames be included?}
}
\value{
A \code{list} of length 2
  \item{data}{A \code{data.frame} containing the columns that were not removed}
  \item{script}{A line of code that can be used to replicate cleaning performed in the gadget}
}
\description{
Shiny gadget used to visually inspect column values in a data set 
and subset rows by specifying column values
}
\examples{
\dontrun{clean_rows(mtcars)}

}
\seealso{
Other shinygadgets: \code{\link{clean_columns}}
}
