% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{create_logo}
\alias{create_logo}
\title{Create a branding logo for a teachingApp}
\usage{
create_logo(app_dir, git_user = NULL, app_pkg = NULL, icon = NULL,
  img = NULL)
}
\arguments{
\item{app_dir}{\code{character} Name of the app in which the signature will be placed}

\item{git_user}{\code{character} GitHub account username (typically created using teachingApps::setOptions())}

\item{app_pkg}{\code{character} Name of package in which \code{app_name} exists}

\item{icon}{\code{character} Name of a fontAwesome icon printed in the app footer}

\item{img}{\code{character} Path to an image printed in the app footer}
}
\value{
HTML code for inserting an icon or image logo in the footer of a navbarPage app
}
\description{
Create a branding logo for a teachingApp
}
