% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degradation_failure.R
\name{degradation_failure}
\alias{degradation_failure}
\title{Function Title}
\usage{
degradation_failure(theme = "flatly", width = "100\%", storyteller = F,
  css = NULL, height = if (storyteller) "800px" else "600px",
  more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{width}{code{character} Width of the printed app}

\item{storyteller}{code{logical} Is this a storyteller app?}

\item{css}{code{character} Path to a custom css file. If code{NULL} the default css file is used}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\description{
Description
}
