% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_lognormal.R
\name{distribution_lognormal}
\alias{distribution_lognormal}
\title{Function Title}
\usage{
distribution_lognormal(theme = "flatly", storyteller = F, width = "100\%",
  height = if (storyteller) "800px" else "600px", more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{storyteller}{code{logical} Is this a storyteller app?}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Description
}
\examples{
\dontrun{
appy(theme = 'spacelab', storyteller = T, pub = F)
}
}
\seealso{
Other distribution apps: \code{\link{distribution_beta}},
  \code{\link{distribution_binomial}},
  \code{\link{distribution_bisa}},
  \code{\link{distribution_egeng}},
  \code{\link{distribution_exponential}},
  \code{\link{distribution_gamma}},
  \code{\link{distribution_geng}},
  \code{\link{distribution_goma}},
  \code{\link{distribution_hypergeometric}},
  \code{\link{distribution_igau}},
  \code{\link{distribution_levgets}},
  \code{\link{distribution_lev}},
  \code{\link{distribution_logistic}},
  \code{\link{distribution_norgets}},
  \code{\link{distribution_normal}},
  \code{\link{distribution_poisson}},
  \code{\link{distribution_sevgets}},
  \code{\link{distribution_sev}},
  \code{\link{distribution_weibull}}

Other htmlwidget apps: \code{\link{acceptance_mtbf_mini}},
  \code{\link{acceptance_mtbf}},
  \code{\link{acceptance_proportion}},
  \code{\link{acceptance_testing_mtbf}},
  \code{\link{complex_ideas}},
  \code{\link{distribution_beta}},
  \code{\link{distribution_binomial}},
  \code{\link{distribution_bisa}},
  \code{\link{distribution_egeng}},
  \code{\link{distribution_exponential}},
  \code{\link{distribution_gamma}},
  \code{\link{distribution_geng}},
  \code{\link{distribution_goma}},
  \code{\link{distribution_hypergeometric}},
  \code{\link{distribution_igau}},
  \code{\link{distribution_levgets}},
  \code{\link{distribution_lev}},
  \code{\link{distribution_logistic}},
  \code{\link{distribution_norgets}},
  \code{\link{distribution_normal}},
  \code{\link{distribution_poisson}},
  \code{\link{distribution_sevgets}},
  \code{\link{distribution_sev}},
  \code{\link{distribution_weibull}}
}
