% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-egeng.R
\name{degengl}
\alias{degengl}
\alias{pegeng}
\alias{qegengl}
\title{The Extended Generalized Gamma Distribution}
\usage{
degengl(x, mu, sigma, delta)

pegeng(q, mu, sigma, delta)

qegengl(p, mu, sigma, delta, smalldelta = 1e-04)
}
\arguments{
\item{x}{A numeric vector of observed values}

\item{mu}{The location parameter}

\item{sigma}{The scale parameter}

\item{delta}{The delta parameter}

\item{q}{A numeric vector of quantile values}

\item{p}{A numeric vector of probability values}

\item{smalldelta}{A numeric shift value}
}
\description{
The Extended Generalized Gamma Distribution
}
