% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure7_3.R
\name{figure7_3}
\alias{figure7_3}
\title{Function Title}
\usage{
figure7_3(theme = "flatly", storyteller = F, width = "100\%",
  height = if (storyteller) "800px" else "600px", css = NULL,
  more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{storyteller}{code{logical} Is this a storyteller app?}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{css}{code{character} Path to a custom css file. If code{NULL} the default css file is used}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\description{
Description
}
