% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-geng.R
\name{dgeng}
\alias{dgeng}
\alias{pgeng}
\alias{qgeng}
\alias{rgeng}
\title{The Generalized Gamma Distribution}
\usage{
dgeng(x, theta, delta, varrho)

pgeng(q, theta, delta, varrho)

qgeng(p, theta, delta, varrho, smalldelta = 1e-05)

rgeng(n, theta, delta, varrho)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{theta}{The theta parameter}

\item{delta}{The delta parameter}

\item{varrho}{The varrho parameter}

\item{q}{A numeric vector of quantiles}

\item{p}{A numeric vector of probabilities}

\item{smalldelta}{A numeric shift value}

\item{n}{The number of random observations}
}
\description{
The Generalized Gamma Distribution
}
