% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPackage.R
\name{getPackage}
\alias{getPackage}
\title{Installs and/or loads an R package}
\usage{
getPackage(pkg = NULL, repo = NULL)
}
\arguments{
\item{pkg}{Name of a package to be installed/loaded (as a character string)}

\item{repo}{The repository from which the package should be installed.}
}
\description{
Installs and/or loads an R package
}
\details{
If \code{repo = NULL} the package will be installed from the CRAN.  Otherwise, \code{repo} is a character string that referring to the GitHub account in which the package is located
}

