% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-gets.R
\name{dgets}
\alias{dgets}
\alias{pgets}
\alias{qgets}
\alias{sgets}
\title{The Generalized Threshold Distribution}
\usage{
dgets(x, alpha, sigma, varzeta, distribution, smallsigma = 2e-05)

pgets(q, alpha, sigma, varzeta, distribution, smallsigma = 2e-05)

qgets(p, alpha, sigma, varzeta, distribution, smallsigma = 1e-05)

sgets(x, alpha, sigma, varzeta, distribution, smallsigma = 2e-05)
}
\arguments{
\item{x}{The x}

\item{alpha}{The alpha}

\item{sigma}{The sigma}

\item{varzeta}{The varzeta}

\item{distribution}{The distribution on which the gets values are based
Either 'normal', 'lev', or 'sev'}

\item{smallsigma}{The small sigma value}

\item{q}{The q}

\item{p}{The p}
}
\description{
Compute values for members of the generalized threshold 
             distribution family. Members include distributions based
             on the normal distribution ("nor-gets"), the smallest-
             extreme value distribution ("sev-gets"), and the largest-
             extreme value distribution ("lev-gets").
}
