% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-goma.R
\name{dgoma}
\alias{dgoma}
\alias{pgoma}
\alias{qgoma}
\title{The Gompertz Makeham Distribution}
\usage{
dgoma(x, shape, shape2, scale = 1)

pgoma(q, shape, shape2, scale = 1)

qgoma(
  p,
  shape = stop("no shape arg"),
  shape2 = stop("no shape2 arg"),
  scale = 1
)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{shape}{A shape parameter}

\item{shape2}{Another shape parameter}

\item{scale}{The scale parameter}

\item{q}{A numeric vector of quantiles}

\item{p}{A numeric vector of probabilities}
}
\description{
The Gompertz Makeham Distribution
}
