% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installApp.R
\name{includeApp}
\alias{includeApp}
\title{Include Shiny apps in documents and presentations}
\usage{
includeApp(appName, appUrl = NULL)
}
\arguments{
\item{appName}{The name of the app, as a character string}

\item{appUrl}{The prefix of the app URL, as a character string (see \code{details})}
}
\description{
This function is a wrapper for \code{knitr::include_app} adding the app's URL prefix
}
\details{
The app prefix is often of the form 'https://accountname.shinyapps.io'
}

