% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/includeR.R
\name{includeArgs}
\alias{includeArgs}
\title{Function Title}
\usage{
includeArgs(path)
}
\arguments{
\item{path}{file path}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, set code{pub = TRUE} to prevent calls to code{install.packages}. Calls to code{install.packages} should not be included within an app and will result in an error.
}

