% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServer.R
\name{insertRmd}
\alias{insertRmd}
\title{Insert an rmarkdown file within a storyteller app}
\usage{
insertRmd(rmd)
}
\arguments{
\item{rmd}{An rmarkdown file saved in the app directory}
}
\description{
Insert an rmarkdown file within a storyteller app
}

