% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServer.R
\name{insertServer}
\alias{insertServer}
\title{Insert an app server into a storyteller app}
\usage{
insertServer(appName, envir = NULL, pkg = "teachingApps")
}
\arguments{
\item{appName}{The app from which the UI is pulled}

\item{pkg}{The package in which the \code{appName} exists}
}
\description{
Insert an app server into a storyteller app
}

