% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServer.R
\name{insertUI}
\alias{insertUI}
\title{Insert an app UI into a storyteller app}
\usage{
insertUI(appName, pkg = "teachingApps")
}
\arguments{
\item{appName}{The app from which the UI is pulled}

\item{pkg}{The package in which the \code{appName} exists}
}
\description{
Insert an app UI into a storyteller app
}

