% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution-igau.R
\name{digau}
\alias{digau}
\alias{pigau}
\alias{qigau}
\title{The Inverse Gaussian Distribution}
\usage{
digau(x, shape, scale = 1)

pigau(q, shape, scale = 1)

qigau(p, shape = stop("no shape arg"), scale = 1)
}
\arguments{
\item{x}{A numeric vector of observations}

\item{shape}{The shape parameter}

\item{scale}{The scale parameter}

\item{q}{A numeric vector of quantiles}

\item{p}{A numeric vector of probabilities}
}
\description{
The Inverse Gaussian Distribution
}
