% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_shiny_app.R
\docType{data}
\name{is_shiny_app}
\alias{is_shiny_app}
\title{Determine if a file is in a shiny app directory}
\format{An object of class \code{root_criterion} of length 5.}
\usage{
is_shiny_app
}
\description{
Determine if a file is in a shiny app directory
}
\keyword{datasets}
