% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{jkf.par}
\alias{jkf.par}
\title{Custom \code{par} function}
\usage{
jkf.par(...)
}
\arguments{
\item{...}{Parameter passed to \code{par} in addition to those defined}
}
\description{
Custom \code{par} function
}
