% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qlev}
\alias{qlev}
\alias{plev}
\alias{dlev}
\alias{rlev}
\title{The Largest Extreme Value Distribution}
\usage{
qlev(p, loc = 0, scale = 1)

plev(q, loc = 0, scale = 1)

dlev(x, loc = 0, scale = 1)

rlev(n, loc = 0, scale = 1)
}
\arguments{
\item{p}{Vector of probabilities}

\item{loc}{Location parameter}

\item{scale}{Scale parameter}

\item{x, }{q Vector of quantiles}

\item{n}{Number of observations}
}
\description{
Density, distribution function, quantile function and 
             random generation for the LEV distribution with location
             \code{loc} and scale \code{scale}.
}
