% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistics_performance.R
\name{logistics_performance}
\alias{logistics_performance}
\title{International Logistics Performance Index App}
\usage{
logistics_performance(theme = "flatly", storyteller = T, width = "100\%",
  height = if (storyteller) "800px" else "600px", more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Description
}
\examples{
\dontrun{
appy(theme = 'spacelab', height = '600px')
}
}
\seealso{
Other plotly apps: \code{\link{acceptance_mtbf}}
}
