% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServer.R
\name{nestRmd}
\alias{nestRmd}
\title{Nest an rmarkdown file within an app}
\usage{
nestRmd(rmd)
}
\arguments{
\item{rmd}{An rmarkdown file saved in the app directory}
}
\description{
Nest an rmarkdown file within an app
}

