% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServer.R
\name{nestServer}
\alias{nestServer}
\title{Nest the server of a smaller app within the server of larger app}
\usage{
nestServer(appName, envir = NULL, pkg = "teachingApps")
}
\arguments{
\item{appName}{The name of the "smaller" app from which the server will be pulled}

\item{envir}{The environment in which the call is made typically \code{environment()}}

\item{pkg}{The package in which \code{appName} exists}
}
\value{
An \code{Observer}-class object resulting from evaluating a \code{server.R} file
}
\description{
Sources a \code{server.R} file before parsing and evaluating its contents in a specified environment
}
\details{
This function is only used within a \code{tabPanel} portion of a shiny app.  
\code{nestServer} may be used with apps stored in other packages.  However, apps are assumed 
to be stored in the apps/ directory located at top level of the package.
}
\examples{
\dontrun{

## server.R from teachingApps::maximum_likelihood

server = function(input, output, session) {

output$mleback1 <- renderUI({ teachingApps::nestRmd('background1.Rmd') })

output$example1_1 <- renderUI({ teachingApps::nestRmd('example1_1.Rmd') })
  teachingApps::nestServer('likelihood_ace', envir = environment())
output$example1_2 <- renderUI({ teachingApps::nestRmd('example1_2.Rmd') })
  teachingApps::nestServer('likelihood_ace2', envir = environment())
output$example1_3 <- renderUI({ teachingApps::nestRmd('example1_3.Rmd') })


output$example2_1 <- renderUI({ teachingApps::nestRmd('example2_1.Rmd') })
  teachingApps::nestServer('exp_mle', envir = environment())
output$example2_2 <- renderUI({ teachingApps::nestRmd('example2_2.Rmd') })
  teachingApps::nestServer('exp_numerical', envir = environment())
output$example2_3 <- renderUI({ teachingApps::nestRmd('example2_3.Rmd') })
  teachingApps::nestServer('soln_numerical2', envir = environment())

  
teachingApps:::nestServer('likelihood_simulation', envir = environment())


output$details <- renderUI({ teachingApps::nestRmd('details.Rmd') })
}
}
}
\seealso{
\code{\link{nestUI} \link{nestRmd}}
}
