% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServer.R
\name{nestUI}
\alias{nestUI}
\title{Nest the UI of smaller app within the UI of a larger app}
\usage{
nestUI(appName, pkg = "teachingApps")
}
\arguments{
\item{appName}{The name of "smaller" app from which the UI will be pulled}

\item{pkg}{The package in which \code{appName} exists}
}
\description{
Nest the UI of smaller app within the UI of a larger app
}

