% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newApp.R
\name{newApp}
\alias{newApp}
\title{Create a new teachingApp}
\usage{
newApp(appName, pkg = "teachingApps", pkg.path = NULL, storyteller = F)
}
\arguments{
\item{appName}{The name of the new teachingApp}

\item{pkg}{Package in which this app will reside (defaults to 'teachingApps')}

\item{pkg.path}{Absolute path to DESCRIPTION file for the package in which the app will reside}

\item{storyteller}{Will this be a storyteller teachingApp?}
}
\description{
Wrapper function to create a new teachingApp
}
