% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_demo.R
\name{pdf_demo}
\alias{pdf_demo}
\title{Function Title}
\usage{
pdf_demo(theme = "flatly", width = "100\%", height = "800px",
  more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Description
}
\examples{
\dontrun{
appy(theme = 'spacelab', height = '600px')
}
}
