% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_positions.R
\name{plotting_positions}
\alias{plotting_positions}
\title{Function Title}
\usage{
plotting_positions(rmd = TRUE, pub = FALSE)
}
\arguments{
\item{rmd}{Will this app be included in an Rmarkdown document or presentation? (see details)}

\item{pub}{Will this app be published? (see details)}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, set \code{pub = TRUE} to prevent calls to \code{install.packages}. Calls to \code{install.packages} should not be included within an app and will result in an error.
}

