% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability_plotting.R
\name{probability_plotting}
\alias{probability_plotting}
\title{Function Title}
\usage{
probability_plotting(theme = "flatly", storyteller = F, width = "100\%",
  height = if (storyteller) "800px" else "600px", more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{storyteller}{code{logical} Is this a storyteller app?}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Description
}
\examples{
\dontrun{
appy(theme = 'spacelab', storyteller = T, pub = F)
}
}
