% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_height.R
\name{set_height}
\alias{set_height}
\title{Dynamically set the height of an app}
\usage{
set_height(height = NULL, dir = NULL)
}
\arguments{
\item{height}{\code{character} Default height of an app (in px)}

\item{dir}{\code{character} App directory path}
}
\description{
Dynamically set the height of an app
}
