% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOptions.R
\name{setOptions}
\alias{setOptions}
\title{Save a local package directory as an environment variable}
\usage{
setOptions(local.dir = NULL, pkg = NULL, git.user = NULL)
}
\arguments{
\item{local.dir}{Path to the local directory at which \code{pkg} is located}

\item{pkg}{Name of a package}

\item{git.user}{The GitHub username}
}
\description{
Save a local package directory as an environment variable
}
