% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qsev}
\alias{qsev}
\alias{psev}
\alias{dsev}
\alias{rsev}
\title{The Smallest Extreme Value Distribution}
\usage{
qsev(p, loc = 0, scale = 1)

psev(q, loc = 0, scale = 1)

dsev(x, loc = 0, scale = 1)

rsev(n, loc = 0, scale = 1)
}
\arguments{
\item{p}{Vector of probabilities}

\item{loc}{Location parameter}

\item{scale}{Scale parameter}

\item{x, }{q Vector of quantiles}

\item{n}{Number of observations}
}
\description{
Density, distribution function, quantile function and 
             random generation for the SEV distribution with location
             \code{loc} and scale \code{scale}.
}
