% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachingApp.R
\name{signature}
\alias{signature}
\title{Place a signature on a teachingApp}
\usage{
signature(app.name, git.username = NULL, appPackage = NULL, icon = NULL,
  img = NULL)
}
\arguments{
\item{app.name}{\code{character} Name of the app in which the signature will be placed}

\item{git.username}{\code{character} GitHub account username (typically created using teachingApps::setOptions())}

\item{appPackage}{\code{character} Name of package in which \code{app.name} exists}
}
\description{
Place a signature on a teachingApp
}
