% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{spmlgeng}
\alias{spmlgeng}
\title{R interface for gng log(1-cdf)}
\usage{
spmlgeng(tvec, gamme, maxlen, answer)
}
\arguments{
\item{tvec}{A numeric vector of observations}

\item{gamme}{A numeric matrix containing the parameter values}

\item{maxlen}{The number of columns in \code{gamme}}

\item{answer}{A numeric vector containing the return values}
}
\description{
R interface for gng log(1-cdf)
}
