% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/superalloy_data17.R
\name{superalloy_data17}
\alias{superalloy_data17}
\title{Function Title}
\usage{
superalloy_data17(pub = FALSE)
}
\arguments{
\item{pub}{Will this app be published? (see details)}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, set \code{pub = TRUE} to prevent calls to \code{install.packages}. Calls to \code{install.packages} should not be included within an app and will result in an error.
}

