% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3_4.R
\name{table3_4}
\alias{table3_4}
\title{Function Title}
\usage{
table3_4(pub = FALSE)
}
\arguments{
\item{pub}{Will this app be published? (see details)}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, set \code{pub = TRUE} to prevent calls to \code{install.packages}. Calls to \code{install.packages} should not be included within an app and will result in an error.
}

