% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table3_5.R
\name{table3_5}
\alias{table3_5}
\title{Function Title}
\usage{
table3_5(theme = "flatly", storyteller = F, width = "100\%", height = if
  (storyteller) "800px" else "600px", more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{storyteller}{code{logical} Is this a storyteller app?}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Description
}
\examples{
\dontrun{
appy(theme = 'spacelab', storyteller = T, pub = F)
}
}
