% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachingApp.R
\name{teachingApp}
\alias{teachingApp}
\title{Render a teachingApp With Options}
\usage{
teachingApp(app_name = NULL, theme = "flatly", width = "100\%",
  height = "800px", stamp = "fa fa-github", more.opts = list(NA), ...)
}
\arguments{
\item{theme}{A \code{character} string naming a bootswatch color theme allowed by \code{shinythemes::shinytheme}}

\item{width}{A \code{character} string specifying the width of the printed app (in pixels)}

\item{height}{A \code{character} string specifying the height of the printed app (in pixels)}

\item{stamp}{A \code{character} string of a fontAwesome icon or path to an image file to be placed in the footer of a navbarPage app}

\item{more.opts}{A \code{list} of additional options/objects that can be passed to the app (see Details)}

\item{...}{A \code{list} of additional options passed to \code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Renders a shiny app from the teachingApps package as a stand-alone app
    or as an element in an rmarkdown document.
}
\details{
The \code{teachingApps} package provides an infrastructure 
   that allows users to dynamically change the appearance and function of
   shiny apps. R users a familiar with writing functions to dynamically alter 
   some output - in this case the output is a app.  Normally,
}
\examples{
dontrun{
teachingApps(app_name = 'distribution_goma', theme = 'spacelab', height = '600px')
}
}
