% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachingApps.R
\docType{package}
\name{teachingApps-package}
\alias{teachingApps}
\alias{teachingApps-package}
\title{teachingApps: Build, Update, and Share Shiny Apps - Quickly and Easily}
\description{
Instantly create shiny app templates and app calling functions
for use in R packages.  Manage large collections of shiny apps from one
location.  Update the structure or appearance of apps on the fly.  Nest
small apps together to create complex 'storyteller' apps. Publish stand-
alone apps or embed in rmarkdown documents with app calling functions.
Includes over 100 shiny apps and shiny gadgets for analyzing data and
teaching various concepts (statistics and R programming).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Auburngrads/teachingApps}
  \item Report bugs at \url{https://github.com/Auburngrads/teachingApps/issues}
}

}
\author{
\strong{Maintainer}: Jason Freels \email{auburngrads@live.com}

Other contributors:
\itemize{
  \item Bradley Boehmke \email{bradleyboehmke@gmail.com} [contributor]
}

}
\keyword{internal}
