% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeApp_dir_dir.R
\name{tradeApp_dir_dir}
\alias{tradeApp_dir_dir}
\title{Transfer a teachingApp app between local directories}
\usage{
tradeApp_dir_dir(app, dir_from = "teachingApps", dir_to = NULL,
  remove = F)
}
\arguments{
\item{app}{\code{character} The app name}

\item{dir_from}{\code{character} The local directory from which the app will be traded (see details)}

\item{dir_to}{The directory to which the app will be traded (see details)}

\item{remove}{\code{logical} If \code{TRUE}, the app and calling function will be removed from \code{from}.}
}
\description{
Transfer a teachingApp app between local directories
}
\details{
see ?teachingApps::tradeApp
}
