% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeApp_dir_pkg.R
\name{tradeApp_dir_pkg}
\alias{tradeApp_dir_pkg}
\title{Transfer a teachingApp from a local directory to a package}
\usage{
tradeApp_dir_pkg(app, dir_from = NULL, pkg_to = NULL, remove = remove)
}
\arguments{
\item{app}{\code{character} The app name}

\item{dir_from}{\code{character} The local directory from which the app will be traded (see details)}

\item{pkg_to}{\code{character} The pacakge to which the app will be traded (see details)}

\item{remove}{\code{logical} If \code{TRUE}, the app and calling function will be removed from \code{from}.}
}
\description{
Transfer a teachingApp from a local directory to a package
}
\details{
see ?teachingApps::tradeApp
}
