% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeApp_pkg_dir.R
\name{tradeApp_pkg_dir}
\alias{tradeApp_pkg_dir}
\title{Transfer a teachingApp from a package to a local directory}
\usage{
tradeApp_pkg_dir(app, pkg_from = "teachingApps", dir_to = NULL,
  remove = remove)
}
\arguments{
\item{app}{\code{character} The app name}

\item{pkg_from}{\code{character} The package from which the app will be traded (see details)}

\item{dir_to}{\code{character} The directory to which the app will be traded (see details)}

\item{remove}{\code{logical} If \code{TRUE}, the app and calling function will be removed from \code{from}.}
}
\description{
Transfer a teachingApp from a package to a local directory
}
\details{
see ?teachingApps::tradeApp
}
