% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tradeApp_pkg_pkg.R
\name{tradeApp_pkg_pkg}
\alias{tradeApp_pkg_pkg}
\title{Transfer a teachingApp between packages}
\usage{
tradeApp_pkg_pkg(app, pkg_from = "teachingApps", pkg_to = NULL,
  remove = remove)
}
\arguments{
\item{app}{\code{character} The app name}

\item{pkg_from}{\code{character} The package from which the app will be traded (see details)}

\item{pkg_to}{\code{character} The pacakge to which the app will be traded (see details)}

\item{remove}{\code{logical} If \code{TRUE}, the app and calling function will be removed from \code{from}.}
}
\description{
Transfer a teachingApp between packages
}
\details{
see ?teachingApps::tradeApp
}
