% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/what_are_you.R
\name{what_are_you}
\alias{what_are_you}
\title{Determine what a thing is}
\usage{
what_are_you(thing)
}
\arguments{
\item{thing}{\code{character} A package name, directory, or file path}
}
\description{
Determine what a thing is
}
