% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acceptance_mtbf.R
\name{acceptance_mtbf}
\alias{acceptance_mtbf}
\title{Reliability acceptance test planning app}
\usage{
acceptance_mtbf(theme = "flatly", width = "100\%", storyteller = F,
  height = if (storyteller) "800px" else "600px", css = NULL,
  more.opts = list(NA), ...)
}
\arguments{
\item{theme}{Character string naming a color theme bootswatch color theme. Must be one of the themes that can be used in \code{shinythemes::shinytheme()}}

\item{width}{Width of the printed app}

\item{storyteller}{Is this a storyteller app?}

\item{height}{Height of the printed app}

\item{css}{Path to a custom css file. If \code{NULL} the default css file is used}

\item{more.opts}{A list of additional options/objects to be passed to the app (see Details)}

\item{...}{Additional options passed to \code{shiny::shinyAppDir()}}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, setting code{pub = TRUE} prevents calls to code{install.packages}. Calls to code{install.packages} should not be included within an app and will result in an error.
}
\seealso{
Other htmlwidget apps: \code{\link{acceptance_mtbf_mini}},
  \code{\link{acceptance_proportion}},
  \code{\link{acceptance_testing_mtbf}},
  \code{\link{complex_ideas}},
  \code{\link{distribution_beta_full}},
  \code{\link{distribution_beta}},
  \code{\link{distribution_binomial}},
  \code{\link{distribution_bisa}},
  \code{\link{distribution_egeng}},
  \code{\link{distribution_exponential_full}},
  \code{\link{distribution_exponential}},
  \code{\link{distribution_gamma}},
  \code{\link{distribution_geng}},
  \code{\link{distribution_goma}},
  \code{\link{distribution_hypergeometric}},
  \code{\link{distribution_igau}},
  \code{\link{distribution_levgets}},
  \code{\link{distribution_lev}},
  \code{\link{distribution_logistic_full}},
  \code{\link{distribution_logistic}},
  \code{\link{distribution_loglogistic_full}},
  \code{\link{distribution_lognormal_full}},
  \code{\link{distribution_lognormal}},
  \code{\link{distribution_norgets}},
  \code{\link{distribution_normal_full}},
  \code{\link{distribution_normal}},
  \code{\link{distribution_poisson_full}},
  \code{\link{distribution_poisson}},
  \code{\link{distribution_sev_full}},
  \code{\link{distribution_sevgets}},
  \code{\link{distribution_sev}},
  \code{\link{distribution_weibull_full}},
  \code{\link{distribution_weibull}}

Other storyteller apps: \code{\link{acceptance_proportion}},
  \code{\link{cdf_demo}}, \code{\link{complex_ideas}},
  \code{\link{distribution_beta_full}},
  \code{\link{distribution_exponential_full}},
  \code{\link{distribution_logistic_full}},
  \code{\link{distribution_loglogistic_full}},
  \code{\link{distribution_lognormal_full}},
  \code{\link{distribution_normal_full}},
  \code{\link{distribution_poisson_full}},
  \code{\link{distribution_sev_full}},
  \code{\link{distribution_weibull_full}}
}

