% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_very_normal.R
\name{ace_very_normal}
\alias{ace_very_normal}
\title{Function Title}
\usage{
ace_very_normal(pub = FALSE)
}
\arguments{
\item{pub}{Will this app be published? (see details)}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, set \code{pub = TRUE} to prevent calls to \code{install.packages}. Calls to \code{install.packages} should not be included within an app and will result in an error.
}

