% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ace_very_normal.R
\name{ace_very_normal}
\alias{ace_very_normal}
\title{Function Title}
\usage{
ace_very_normal(theme = "flatly", width = "100\%", storyteller = F,
  height = if (storyteller) "800px" else "600px", css = NULL,
  more.opts = list(NA), ...)
}
\arguments{
\item{theme}{Character string naming a color theme bootswatch color theme. Must be one of the themes that can be used in code{shinythemes::shinytheme()}}

\item{width}{Width of the printed app}

\item{storyteller}{Is this a storyteller app?}

\item{height}{Height of the printed app}

\item{css}{Path to a custom css file. If code{NULL} the default css file is used}

\item{more.opts}{A list of additional options/objects to be passed to the app (see Details)}

\item{...}{Additional options passed to \code{shiny::shinyAppDir()}}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, setting code{pub = TRUE} prevents calls to code{install.packages}. Calls to code{install.packages} should not be included within an app and will result in an error.
}
\seealso{
Other shinyAce apps: \code{\link{basic_histogram}},
  \code{\link{cdf_demo}}
}
