% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_css.R
\name{add_css}
\alias{add_css}
\title{Add CSS style rules to an teachingApp}
\usage{
add_css()
}
\value{
teachingApps style rules defined to an app
}
\description{
Add CSS style rules to an teachingApp
}
\seealso{
code{link{add_theme}}
}
