% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_css.R
\name{add_css}
\alias{add_css}
\title{Add teachingApps CSS style rules to an app}
\usage{
add_css()
}
\value{
teachingApps style rules defined to an app
}
\description{
Add teachingApps CSS style rules to an app
}
\details{
This function should not be called directly but must be 
         be included within the body of an app's \code{ui}
}
\seealso{
\code{\link{add_theme}}
}
