% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_rmd.R
\name{add_rmd}
\alias{add_rmd}
\title{Add an rmarkdown file to an app}
\usage{
add_rmd(rmd, path)
}
\arguments{
\item{rmd}{\code{character} Name of an rmarkdown file saved in the app directory}

\item{path}{Path to a file outside of the app directory}
}
\description{
Run inline and stand-alone code chunks and include results
as part of a shiny app. Include \code{LaTeX}-typeset equations with MathJax
}
\examples{
\dontrun{
# see examples in add_server and add_ui documentation
}
}
\seealso{
\code{\link{add_server} \link{add_ui}}
}
