% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_css.R
\name{add_theme}
\alias{add_theme}
\title{Dynamically add css rules to an app}
\usage{
add_theme(theme = NULL)
}
\arguments{
\item{main.css}{The main CSS file}

\item{user.css}{A user specified CSS file}

\item{name}{The name of the app}
}
\description{
Dynamically add css rules to an app
}
