% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bent_probplot.R
\name{bent_probplot}
\alias{bent_probplot}
\title{SMRD non-linear probability plot app}
\usage{
bent_probplot(theme = "flatly", width = "100\%", storyteller = F,
  css = NULL, height = if (storyteller) "800px" else "600px",
  more.opts = list(NA), ...)
}
\arguments{
\item{theme}{Character string naming a color theme bootswatch color theme. Must be one of the themes that can be used in code{shinythemes::shinytheme()}}

\item{width}{Width of the printed app}

\item{storyteller}{Is this a storyteller app?}

\item{css}{Path to a custom css file. If code{NULL} the default css file is used}

\item{height}{Height of the printed app}

\item{more.opts}{A list of additional options/objects to be passed to the app (see Details)}

\item{...}{Additional options passed to \code{shiny::shinyAppDir()}}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, setting code{pub = TRUE} prevents calls to code{install.packages}. Calls to code{install.packages} should not be included within an app and will result in an error.
}

