% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{qbisa}
\alias{qbisa}
\alias{pbisa}
\alias{dbisa}
\alias{rbisa}
\title{The Birmbaum-Saunders Distribution}
\usage{
qbisa(p, shape, scale = 1)

pbisa(q, shape, scale = 1)

dbisa(x, shape, scale = 1)

rbisa(n, shape, scale = 1)
}
\arguments{
\item{p}{Vector of probabilities}

\item{shape}{Shape parameter}

\item{scale}{Scale parameter}

\item{q}{Vector of quantiles}

\item{x}{Vector of quantiles}

\item{n}{Number of observations}
}
\description{
Density, distribution function, quantile function and 
             random generation for the BISA distribution with location
             \code{loc} and scale \code{scale}.
}
