% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cdf_demo.R
\name{cdf_demo}
\alias{cdf_demo}
\title{Visualization of Cumulative Distribution Function}
\usage{
cdf_demo(theme = "flatly", storyteller = F, width = "100\%", height = if
  (storyteller) "800px" else "600px", more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{storyteller}{code{logical} Is this a storyteller app?}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Description
}
\examples{
\dontrun{
appy(theme = 'spacelab', storyteller = T, pub = F)
}
}
\seealso{
Other shinyAce apps: \code{\link{ace_very_normal}},
  \code{\link{basic_histogram}}

Other storyteller apps: \code{\link{acceptance_mtbf}},
  \code{\link{acceptance_proportion}},
  \code{\link{complex_ideas}}
}
