% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_beta.R
\name{distribution_beta}
\alias{distribution_beta}
\title{Visualization app for the beta distribution}
\usage{
distribution_beta(theme = "flatly", width = "100\%", height = "800px",
  more.opts = list(NA), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Description
}
\examples{
\dontrun{
appy(theme = 'spacelab', height = '600px')
}
}
\seealso{
Other distribution apps: \code{\link{distribution_binomial}},
  \code{\link{distribution_bisa}},
  \code{\link{distribution_egeng}},
  \code{\link{distribution_exponential}},
  \code{\link{distribution_gamma}},
  \code{\link{distribution_geng}},
  \code{\link{distribution_goma}},
  \code{\link{distribution_hypergeometric}},
  \code{\link{distribution_igau}},
  \code{\link{distribution_levgets}},
  \code{\link{distribution_lev}},
  \code{\link{distribution_logistic}},
  \code{\link{distribution_lognormal}},
  \code{\link{distribution_norgets}},
  \code{\link{distribution_normal}},
  \code{\link{distribution_poisson}},
  \code{\link{distribution_sevgets}},
  \code{\link{distribution_sev}},
  \code{\link{distribution_weibull}}

Other htmlwidget apps: \code{\link{acceptance_mtbf_mini}},
  \code{\link{acceptance_proportion}},
  \code{\link{acceptance_testing_mtbf}},
  \code{\link{complex_ideas}},
  \code{\link{distribution_binomial}},
  \code{\link{distribution_bisa}},
  \code{\link{distribution_egeng}},
  \code{\link{distribution_exponential}},
  \code{\link{distribution_gamma}},
  \code{\link{distribution_geng}},
  \code{\link{distribution_goma}},
  \code{\link{distribution_hypergeometric}},
  \code{\link{distribution_igau}},
  \code{\link{distribution_levgets}},
  \code{\link{distribution_lev}},
  \code{\link{distribution_logistic}},
  \code{\link{distribution_lognormal}},
  \code{\link{distribution_norgets}},
  \code{\link{distribution_normal}},
  \code{\link{distribution_poisson}},
  \code{\link{distribution_sevgets}},
  \code{\link{distribution_sev}},
  \code{\link{distribution_weibull}}
}
