% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribution_loglogistic.R
\name{distribution_loglogistic}
\alias{distribution_loglogistic}
\title{Function Title}
\usage{
distribution_loglogistic(pub = FALSE, theme = "flatly", rmd = FALSE, ...)
}
\arguments{
\item{pub}{Will this app be published? (see details)}

\item{theme}{Character string naming a color theme bootswatch color theme. Must be one of the themes that can be used in code{shinythemes::shinytheme()}}

\item{rmd}{Will this be included in a Rmarkdown document or presentation?}

\item{...}{Additional arguments for code{shiny::runApp()}}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, set code{pub = TRUE} to prevent calls to code{install.packages}. Calls to code{install.packages} should not be included within an app and will result in an error.
}

