% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_lm.R
\name{gadget_lm}
\alias{gadget_lm}
\title{Function Title}
\usage{
gadget_lm(data, xvar, yvar, theme = "flatly", width = "100\%", css = NULL,
  height = "600px", ...)
}
\arguments{
\item{data}{A \code{data.frame} object}

\item{xvar}{Column from \code{data} to display on the x-axis}

\item{yvar}{Column from \code{data} to display on the y-axis}

\item{theme}{Character string naming a color theme bootswatch color theme. Must be one of the themes that can be used in code{shinythemes::shinytheme()}}

\item{width}{Width of the printed app.}

\item{css}{Path to a custom css file. If code{NULL} the default css file is used}

\item{height}{Height of the printed app.}

\item{...}{Additional arguments passed to code{shiny::runApp()}}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, setting code{pub = TRUE} prevents calls to code{install.packages}. Calls to code{install.packages} should not be included within an app and will result in an error.
}

