% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood_ace.R
\name{likelihood_ace}
\alias{likelihood_ace}
\title{Function Title}
\usage{
likelihood_ace(theme = "flatly", storyteller = F, width = "100\%",
  height = if (storyteller) "800px" else "600px", more.opts = list(time =
  Sys.time()), ...)
}
\arguments{
\item{theme}{code{character} A bootswatch color theme for use in code{shinythemes::shinytheme}}

\item{width}{code{character} Width of the printed app}

\item{height}{code{character} Height of the printed app}

\item{more.opts}{code{list} Additional options/objects passed to the app (see Details)}

\item{...}{code{list} Additional options passed to code{shiny::shinyAppDir()}}
}
\value{
A printed shiny app
}
\description{
Description
}
\examples{
\dontrun{
appy(theme = 'spacelab', height = '600px')
}
}
