% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maximum_likelihood.R
\name{maximum_likelihood}
\alias{maximum_likelihood}
\title{Function Title}
\usage{
maximum_likelihood(theme = "flatly", width = "100\%", height = if
  (storyteller) "800px" else "600px", storyteller = T, css = NULL,
  more.opts = list(time = Sys.time()), ...)
}
\arguments{
\item{theme}{Character string naming a color theme bootswatch color theme. Must be one of the themes that can be used in \code{shinythemes::shinytheme()}}

\item{width}{Width of the printed app}

\item{height}{Height of the printed app}

\item{storyteller}{Is this a storyteller app?}

\item{css}{Path to a custom css file. If \code{NULL} the default css file is used}

\item{more.opts}{A list of additional options/objects to be passed to the app (see Details)}

\item{...}{Additional options passed to \code{shiny::shinyAppDir()}}
}
\description{
Description
}
\details{
When publishing apps using shinyapps.io or shinyServer, setting code{pub = TRUE} prevents calls to code{install.packages}. Calls to code{install.packages} should not be included within an app and will result in an error.
}

