% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appServer.R
\name{nestServer}
\alias{nestServer}
\title{Nest the server of a smaller app within the server of larger app}
\usage{
nestServer(appName, envir = NULL, pkg = "teachingApps")
}
\arguments{
\item{appName}{The name of the "smaller" app from which the server will be pulled}

\item{envir}{The environment in which the call is made typically \code{environment()}}

\item{pkg}{The package in which \code{appName} exists}
}
\description{
Nest the server of a smaller app within the server of larger app
}

