% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachingApp.R
\name{teachingApp}
\alias{teachingApp}
\title{Leave signature for a teachingApp}
\usage{
teachingApp(app.name, git.username = NULL, appPackage = NULL)
}
\arguments{
\item{app.name}{Name of the app \code{character}}

\item{git.username}{GitHub username (typically created using teachingApps::setOptions())}

\item{appPackage}{Name of package in which \code{app.name} exists}
}
\description{
Leave signature for a teachingApp
}

