% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{teachingApp}
\alias{teachingApp}
\title{Leave signature for teachingApps}
\usage{
teachingApp(name)
}
\arguments{
\item{name}{The name of the app \code{character}}
}
\description{
Leave signature for teachingApps
}

