% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachingApps.R
\docType{package}
\name{teachingApps-package}
\alias{teachingApps}
\alias{teachingApps-package}
\title{teachingApps: Apps for Teaching Statistics, R Programming, and Shiny App Development}
\description{
Contains apps and gadgets for teaching data analysis and
statistics concepts along with how to implement them in R.  Includes
tools to make app development easier and faster by nesting apps together.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Auburngrads/teachingApps}
  \item Report bugs at \url{https://github.com/Auburngrads/teachingApps/issues}
}

}
\author{
\strong{Maintainer}: Jason Freels \email{auburngrads@live.com}

Other contributors:
\itemize{
  \item Bradley Boehmke \email{bradleyboehmke@gmail.com} [contributor]
}

}
\keyword{internal}
