\name{teachingApps-package}
\alias{teachingApps-package}
\alias{teachingApps}
\docType{package}
\title{
\packageTitle{teachingApps}
}
\description{
\packageDescription{teachingApps}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{teachingApps}
\packageIndices{teachingApps}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{teachingApps}

Maintainer: \packageMaintainer{teachingApps}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
