% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateApp.R
\name{updateApp}
\alias{updateApp}
\title{Quickly update the code for a shiny app}
\usage{
updateApp(app = NULL, repo = "teachingApps", pkg = repo,
  lib = .libPaths()[1], open.folder = FALSE, css = FALSE,
  create = FALSE, git = NULL)
}
\arguments{
\item{app}{Name of the app to be updated}

\item{repo}{Name of the directory in which the raw app code is stored}

\item{pkg}{Name of the package in which the updates are to be made.  Defaults to \code{repo}.}

\item{lib}{Package library directory where \code{pkg} is stored. Defaults to \code{.libPaths()[1]}}

\item{open.folder}{If \code{TRUE}, \code{browseURL()} is called to open view the files in the app directory}

\item{css}{Logical variable indicating if the css files should be updated}

\item{create}{Logical inidicating if a new app directory should be created}

\item{git}{Local directory in which git repositories are stored (see details)}
}
\description{
Quickly update the code for a shiny app
}
\details{
This function assumes that the raw app code is in the \code{inst/} directory
Prior to using this function the default git directory must be established in \code{options} using \code{teachingApps::setGit()}
}

